+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ### Changes ###
 
-  * [Core]: Fix an issue with WindTools enabled
+  * [Raid Manager]: Some code improvements to RaidManager
+  * [ChatText]: Here also some code improvements
+  * [RaiderIO Skin]: Updated a new SearchFrame
+  * [Details Skin]: Fix gradient Names and enable it by default
+  * [TalentLoadoutEx]: Remove skin
+  * [Event Tracker]: Change the defaults on the Map Event Tracker
+  * [AutoButtons]: Added S4 openable boxes